Moves utility methods to the utility class.
