+++ src/main/java/org/jitsi/jigasi/util/Util.java
@@ -1,7 +1,7 @@
 /*
  * Jigasi, the JItsi GAteway to SIP.
  *
- * Copyright @ 2015 Atlassian Pty Ltd
+ * Copyright @ 2018 - present 8x8, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,12 +23,13 @@ import net.java.sip.communicator.util.*;
 import org.gagravarr.ogg.*;
 import org.gagravarr.opus.*;
 import org.jitsi.impl.neomedia.codec.*;
-import org.jitsi.impl.neomedia.rtp.*;
 import org.jitsi.service.neomedia.*;
 import org.jitsi.service.neomedia.codec.*;
 import org.jitsi.service.neomedia.format.*;
 import org.jitsi.utils.*;
+import org.jivesoftware.smack.bosh.*;
 
+import java.lang.reflect.*;
 import java.util.*;
 
 import javax.xml.transform.*;
@@ -283,4 +284,55 @@ public class Util
         }
     }
 
+    /**
+     * Extracts bosh connection sessionId if possible.
+     * @param connection the bosh connection which sessionId we will try to
+     * extract.
+     * @return the sessionId if extracted or null.
+     */
+    public static Object getConnSessionId(Object connection)
+    {
+        Field myField = getField(XMPPBOSHConnection.class, "sessionID");
+
+        if (myField != null)
+        {
+            myField.setAccessible(true);
+            try
+            {
+                return myField.get(connection);
+            }
+            catch( Exception e)
+            {
+                Logger.getLogger(Util.class).error("cannot read it", e);
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * Utility method to get the field from a class.
+     * @param clazz the class.
+     * @param fieldName the field men.
+     * @return the field or null.
+     */
+    private static Field getField(Class clazz, String fieldName)
+    {
+        try
+        {
+            return clazz.getDeclaredField(fieldName);
+        }
+        catch (NoSuchFieldException e)
+        {
+            Class superClass = clazz.getSuperclass();
+            if (superClass == null)
+            {
+                return null;
+            }
+            else
+            {
+                return getField(superClass, fieldName);
+            }
+        }
+    }
 }
